[add]ユーティリティファイル作成　main.phpは共通処理を書くファイルで制限したいため
