+++ TechpitFF/main.php
@@ -6,6 +6,7 @@
 // require_once('./classes/BlackMage.php');
 // require_once('./classes/WhiteMage.php');
 require_once('./lib/Loader.php');
+require_once('./lib/Utility.php');
 
 //オートロード
 $loader = new Loader();
@@ -43,19 +44,7 @@ $isFinishFlg = false;
 
 $messageObj = new Message;
 
-//終了条件の判定
-function isFinish($objects){
-    $deathCnt = 0; //HPが0以下の仲間の数
-    foreach($objects as $object){
-        if($object->getHitPoint()>0){
-            return false;
-        }
-        $deathCnt++;
-    }
-    if($deathCnt === count($objects)){
-        return true;
-    }
-}
+
 
 while(!$isFinishFlg){
